+++ py_client/constants.py
@@ -8,3 +8,4 @@ DEFAULT_CONFIG = "config.json"
 DEFAULT_HOME_CONFIG = os.path.join(os.path.expanduser("~"), ".braincube", DEFAULT_CONFIG)
 NO_CONFIG_MSG = "The client needs a configuration file."
 DEFAULT_PAGE_SIZE = 150
+DEFAULT_PARSE_DATE = False