+++ tests/test_memory_base.py
@@ -2,6 +2,7 @@
 """Tests for the memory_base module."""
 
 import responses
+from py_client import parameters
 
 from tests.mock import mb_obj, mock_client, mock_request_entity
 
@@ -24,15 +25,19 @@ def test_get_variable(mb_obj, mock_client):
 
 @responses.activate
 def test_get_variable_list(mocker, monkeypatch, mb_obj, mock_request_entity):
-
     mock_request_entity.return_value = {
-        "items": [{"standard": "name{}".format(i), "bcId": str(i)} for i in range(2)]
+        "items": [{"standard": "name{}".format(i), "bcId": str(i)} for i in range(3)]
     }
-    var_list = mb_obj.get_variable_list(page=0, page_size=2)
+    parameters.set_parameter({"page_size": 2})
+    var_list = mb_obj.get_variable_list(page=0)
     assert var_list[0]._name == "name0"
-    mock_request_entity.assert_called_with(
-        "braincube/bcname/braincube/mb/1/variables/summary?offset=0&size=2"
-    )
+
+    request_path = "braincube/bcname/braincube/mb/1/variables/summary?offset=0&size=2"
+    mock_request_entity.assert_called_with(request_path)
+    mock_request_entity.reset_mock()
+    parameters.set_parameter({"page_size": 4})
+    var_list = mb_obj.get_variable_list(page=0, page_size=2)
+    mock_request_entity.assert_called_with(request_path)
 
 
 @responses.activate
@@ -45,8 +50,12 @@ def test_get_job(mb_obj, mock_client):
 
 @responses.activate
 def test_get_job_list(mocker, monkeypatch, mb_obj, mock_request_entity):
-    job_list = mb_obj.get_job_list(page=0, page_size=2)
+    parameters.set_parameter({"page_size": 2})
+    job_list = mb_obj.get_job_list(page=0)
     assert job_list[0]._name == "name0"
-    mock_request_entity.assert_called_with(
-        "braincube/bcname/braincube/mb/1/jobs/all/summary?offset=0&size=2"
-    )
+    request_path = "braincube/bcname/braincube/mb/1/jobs/all/summary?offset=0&size=2"
+    mock_request_entity.assert_called_with(request_path)
+    mock_request_entity.reset_mock()
+    parameters.set_parameter({"page_size": 4})
+    job_list = mb_obj.get_job_list(page=0, page_size=2)
+    mock_request_entity.assert_called_with(request_path)